✨ Front-end: Updated ExploreComponent, InsertComponent, and StructureComponent
- Modified ExploreComponent to include the `id` property when opening the dialog.
- Modified InsertComponent to include the `id` property when opening the dialog.
- Modified StructureComponent to include the `id` property when opening the dialog.

These changes were made to ensure that the correct data is passed to the dialog components when they are opened.

📝 Documentation: Updated README.md

- Removed the `export` keyword from the list of features for MongoDB.
- Added new lines at the end of the file.

These changes were made to improve the readability of the README file and remove unnecessary information.

🔧 UI: Updated logs-dialog.html

- Added a hyperlink to open a new issue in GitLab for reporting bugs.

This change was made to provide users with an easy way to report bugs by directly opening a new issue in the GitLab repository.

💄 UI: Updated top-right.component.ts

- Initialized the `str` variable as an empty object.

This change was made to prevent potential errors in the `LogsDialog` component.