art: Change drawer mode to "over" in container.component.html
The drawer mode in the container component is changed to "over" instead of "side".

:art: Remove unnecessary styles from container.component.scss

Some unnecessary styles are removed from the container component stylesheet.

:art: Remove unnecessary styles from explore.component.scss

Some unnecessary styles are removed from the explore component stylesheet.

:art: Remove unnecessary styles from structure.component.scss

Some unnecessary styles are removed from the structure component stylesheet.

:sparkles: Add expand and delete functionality in relations.component.html

The relations component HTML is updated to add expand and delete functionality. The expand functionality shows and hides additional details of each relation, while the delete functionality allows removing a relation.

:art: Add styles for expand and additional details in relations.component.scss

Styles for the expand functionality and additional details in the relations component are added.

:bug: Fix typo in relations.component.html

A typo in the relations component HTML is fixed. The class name "expended" is updated to "expanded".

:fire: Remove unnecessary styles from tables.component.scss

Some unnecessary styles are removed from the tables component stylesheet.

:art: Update styles for mat-tooltip in styles.scss

Styles for the mat-tooltip component are updated in the global stylesheet.

:recycle: Refactor RelationsComponent to improve code readability

The RelationsComponent class is refactored to improve code readability and maintainability.