art: Refactor connection component code
Refactored the `connection.component.html` and `connection.component.ts` files.
In the HTML file, changed the `matTooltip` property to use the `getDetails` method instead of the `getTooltip` method to display the tooltip text.
In the TS file, modified the `getDetails` method to return an object with the icon and tooltip text based on the server's status.
Used the object returned by the `getDetails` method to display the icon and tooltip text in the HTML file.

The changes were done to improve the readability and maintainability of the code in the connection component.

:recycle: Refactor structure component code

Refactored the `structure.component.html` file.
Changed the `color` attribute value of the delete button to "warn" and changed the mat-icon-button content from "remove" to "delete".

The changes were done to provide a more consistent and descriptive UI in the structure component.

:lipstick: Update cell component UI

Updated the `cell.component.html` file.
Added logic to conditionally display JSON data as either a collapsed or expanded view.
Added a button to toggle the expanded/collapsed view of JSON data.

The changes were done to improve the user experience when working with JSON data in the cell component.

:bug: Fix cell component styling

Fixed the `cell.component.scss` file.
Adjusted the `max-width` and `max-height` values to make the cell component fit within the viewport without overflowing.

The changes were done to improve the display of the cell component and prevent content overflow.

:sparkles: Add column labels to update data dialog

Added column labels to the `update-data-dialog.html` file.
Displayed the column labels using the mat-label directive within a mat-form-field.

The changes were done to provide a more user-friendly interface when updating data in the update data dialog.] update ConnectionComponent and CellComponent

- [front] update ConnectionComponent:
  - Replace function getTooltip with getDetails
  - Modify the return value of getDetails to return an object with icon and txt properties
  - Modify the usage of getTooltip to getDetails in connection.component.html
  - Simplify getIcon function by replacing hardcoded string values with object properties
  - Modify the return value of getIcon to return an object with icon and txt properties
  - Replace hardcoded values in connection.component.html with object properties returned by getDetails and getIcon functions
  - Remove unnecessary imports in connection.component.ts

- [front] update StructureComponent:
  - Replace "remove" icon with "delete" icon in structure.component.html

- [front] update CellComponent:
  - Adjust max-width and max-height properties in cell.component.scss
  - Set expand property to true by default in cell.component.ts
  - Add ElementRef dependency to constructor in cell.component.ts
  - Set resize property of host element to "both" when expand property is false and column value is a large JSON object in cell.component.ts

- [front] update UpdateDataDialogComponent:
  - Display the content of updateSuggestions object only when it's not empty in update-data-dialog.html

- [front] update UpdateDataDialog:
  - Add readonly Object property to allow iteration over object keys in update-data-dialog.ts