🚑 Fix MongoDB trigger validator in replaceTrigger method
The `validator` parameter in the `replaceTrigger` method of the `MongoDB`
wrapper was being passed as a string instead of an object. This commit fixes
the issue by parsing the `trigger.code` as JSON before passing it as the
`validator` value.

This change is necessary to ensure that the MongoDB trigger is correctly
replaced with the new validator code.