recycle: Refactor MongoDB wrapper
Refactor the `replaceTrigger` and `dropTrigger` methods in the `MongoDB` class.
Now, these methods use try-catch blocks to catch and handle any exceptions that may occur when executing the respective MongoDB commands. If an error occurs, the error message is returned as part of the response object.

These changes were made to improve error handling and provide more informative error messages to the users of the MongoDB wrapper.