whale: chore: Update Dockerfile
The Dockerfile was updated to improve the build process for the application.
Previously, the file was copying the `shared-helper.mjs` file before copying the `front` and `back` directories, which caused the build to fail as the `shared-helper.mjs` file was not found in the correct location.

To fix this issue, the `shared-helper.mjs` file is now copied after copying the `front` and `back` directories, ensuring that the file is available in the correct location during the build process.

This change improves the build process for the Docker image by correctly copying the necessary files and preventing build failures.