books: README.md: add TODO public
The commit adds a TODO item to the README.md file indicating that the task "public" needs to be implemented.

:bug: back/src/wrapper/mongodb.js: fix aggregate injection

The commit fixes an issue in the `mongodb.js` file. The code was injecting extra data into the aggregate query, causing an error. The fix ensures that `agg` is an array before pushing `toInject` into it.

:no_entry_sign: front/src/app/container/container.component.ts: remove unused imports

The commit removes two unused imports from the `container.component.ts` file: `MatIconRegistry` and `DomSanitizer`.

:art: front/src/app/tables/tables.component.html: improve code formatting

The commit improves the code formatting in the `tables.component.html` file. It adds line breaks and proper indentation to enhance readability.

:sparkles: front/src/app/tables/tables.component.ts: add width attribute

The commit adds a new `width` attribute to the `TablesComponent` class in the `tables.component.ts` file. The attribute will be used to store the width of the table list sidenav.

:recycle: front/src/app/tables/tables.component.ts: load and save table width

The commit modifies the `TablesComponent` class in the `tables.component.ts` file. It adds functionality to load and save the width of the table list sidenav from local storage. The width is saved under the `tableWidth` key.

:wrench: front/src/shared/cell/cell.component.html: fix expand button behavior

The commit fixes the behavior of the expand button in the `cell.component.html` file. When clicked, the button now sets the `expand` property to `true` and also sets the `resize` style property of the element to `'both'`, allowing the cell to be resized.

:art: front/src/shared/cell/cell.component.scss: improve cell styling

The commit improves the styling of the cell component by modifying the `cell.component.scss` file. It adjusts the `max-width` property to `25vw` and removes unnecessary padding from the `.segment` class in the JSON viewer.

:hammer_and_wrench: front/src/shared/cell/cell.component.ts: add ElementRef to constructor

The commit adds the `ElementRef` parameter to the constructor of the `CellComponent` class in the `cell.component.ts` file. The `ElementRef` will be used to access the native element and modify its style.