books: Add information on unique constraints in README.md
The commit adds information about unique constraints in the README.md file.
The purpose of this change is to provide additional information to users about the unique constraints feature in the application.

:wrench: Add dropView method in MongoDB wrapper

The commit adds a new method called `dropView` to the `MongoDB` wrapper class in the `mongodb.js` file.
The method is responsible for dropping a view/collection from the specified database.
This change was made to provide a convenient way to drop views/collections when needed.

:pencil2: Update styles.scss file

The commit updates the `styles.scss` file in the `front/src` directory.
The changes include adding CSS styles for the `::-webkit-resizer` and `::-webkit-scrollbar-corner` selectors.
These styles enhance the visual appearance and behavior of resizers and scrollbars in the application.

These changes were made to improve the usability and aesthetics of the application's user interface.