🚀 feat: Add database sample functionality to the Controller
Added a new `sample` method to the `Controller` class that retrieves a sample of data from a given database. The method makes use of the `sampleDatabase` method in the database wrapper to fetch the sample data. The sample data includes the structure of tables, data in the tables, and triggers associated with the tables. The method also retrieves server variables if specified. The retrieved sample data is then formatted into a text response and sent back to the client.

This change was made to provide users with a way to get a glimpse of the data present in a database, including the structure of tables and associated triggers. This can be useful for understanding the data model and making informed decisions while working with the database.

The new functionality also includes support for different types of databases such as MongoDB, MySQL, and PostgreSQL.